feat: 출석 테이블 ui 수정
